Some updates
Fleshed out variable names, updated the status effects to read the current status address. Broke out the item and armor arrays to be more human readable. Increased the sizes of the icons for the inventory to 20px versus 15px. Also reworked the entire familiar area and how XP was calculated.