Small update
Updated default tab to 8 spaces to match the GitHub default. Refactored the status to work from an array incase I find more values to add. Also refactored the code for the roomString and how it incorporatesthe currentRoom count and then the calculation for the Room % and using the string.format() function to make it look good.